Fix qwen_image_edit_f2p: cast transformer to target dtype after fuse_lora
LoRA weights load in BFloat16 regardless of the pipeline dtype, leaving
some weights in BFloat16 after fuse_lora() while inputs are Float32.
Explicitly cast the transformer to the target dtype after fusing.

+++ qwen_image_edit_f2p/pytorch/loader.py
@@ -87,7 +87,7 @@ class ModelLoader(ForgeModel):
         )
         pipe.fuse_lora()
 
-        self._transformer = pipe.transformer
+        self._transformer = pipe.transformer.to(dtype)
         self._transformer.eval()
         return self._transformer
